+++ packages/node-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # 0.5.20
 
+## 0.8.11
+
+### Patch Changes
+
+- Up utils
+
 ## 0.8.10
 
 ### Patch Changes